requirements/base.txt
playwright
playwright-stealth
pydantic
jsonpath-ng
pandas
langchain
langchain-openai
langchain-google-genai
jinja2
termcolor
Django>=5.2,<7.0
tendo
pyyaml
fastembed
scikit-learn
huggingface_hub
django-unfold
django-cors-headers
questionary
